fix: resolve all eBPF collector build issues and test failures
Major fixes across 16 collectors to ensure proper compilation:

eBPF Collectors (8):
- kernel: Fixed unused import, type conversion for event_type, added EVENT_TYPE_POD_SYSCALL
- syscall-errors: Moved Description to Custom data, fixed domain type references
- cri-ebpf: Fixed BPF type names (CrimonitorObjects → crimonitorObjects)
- etcd-ebpf: Fixed pointer issues, corrected CollectionConfig field names
- systemd, namespace-collector, oom, storage-io: Added proper Linux build tags

Other Collectors (8):
- Added stub implementations for non-Linux builds
- Fixed domain type mismatches across all collectors
- systemd-api: Added Linux build tags for systemd journal dependencies

Test Fixes:
- namespace-collector: Added EventTypeCNI to domain for proper event typing
- oom: Fixed prediction tolerance, validation messages, struct size comparisons

Infrastructure:
- Created colima-build.sh script for Linux eBPF development on macOS
- Removed 1,315 lines of conflicting stub code
- Consolidated stub implementations with proper build tags

All 16 collectors now build successfully on both Linux and macOS platforms.

+++ pkg/collectors/kernel/types.go
@@ -1,16 +1,18 @@
+//go:build linux
+
 package kernel
 
 // Event type constants - must match C definitions
 const (
-	EventTypeConfigMapAccess = uint8(1) // ConfigMap access events
-	EventTypeSecretAccess    = uint8(2) // Secret access events
-	EventTypePodSyscall      = uint8(3) // Pod syscall events for correlation
+	EventTypeConfigMapAccess = uint32(1) // ConfigMap access events
+	EventTypeSecretAccess    = uint32(2) // Secret access events
+	EventTypePodSyscall      = uint32(3) // Pod syscall events for correlation
 
 	// Legacy event types - kept for test compatibility only
 	// These are no longer monitored but tests still reference them
-	EventTypeProcess = uint8(10) // Deprecated - use syscall-errors collector
-	EventTypeFile    = uint8(11) // Deprecated - covered by ConfigMap/Secret access
-	EventTypeNetwork = uint8(12) // Deprecated - use network collector
+	EventTypeProcess = uint32(10) // Deprecated - use syscall-errors collector
+	EventTypeFile    = uint32(11) // Deprecated - covered by ConfigMap/Secret access
+	EventTypeNetwork = uint32(12) // Deprecated - use network collector
 )
 
 // KernelEvent represents a kernel event from eBPF - must match C struct kernel_event
